% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{as.data.frame.disto}
\alias{as.data.frame.disto}
\title{Convert a disto object to dataframe}
\usage{
\method{as.data.frame}{disto}(x, ...)
}
\arguments{
\item{x}{object of class disto}

\item{...}{arguments for \code{\link[broom]{tidy}}}
}
\value{
a dataframe in long format
}
\description{
Convert the underlying data of a disto object to a dataframe in
  long format (3 columns: item1, item2, distance). This might be a costly
  operation and should be used with caution.
}
\examples{
temp <- stats::dist(iris[,1:4])
dio  <- disto(objectname = "temp")
dio
head(as.data.frame(dio))
}
