% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extractors.R
\name{dist_ij_k}
\alias{dist_ij_k}
\title{Vectorized version of dist_ij_k_}
\usage{
dist_ij_k(i, j, size)
}
\arguments{
\item{i}{row indexes}

\item{j}{column indexes}

\item{size}{value of size attribute of the dist object}
}
\value{
k indexes
}
\description{
Convert ij indexes to k indexes for a dist object
}
