% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extractors.R
\name{dist_ij_k_}
\alias{dist_ij_k_}
\title{Convert ij index to k index}
\usage{
dist_ij_k_(i, j, size)
}
\arguments{
\item{i}{row index}

\item{j}{column index}

\item{size}{value of size attribute of the dist object}
}
\value{
k index
}
\description{
Convert ij index to k index for a dist object
}
