% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extractors.R
\name{dist_k_ij}
\alias{dist_k_ij}
\title{Vectorized version of dist_k_ij_}
\usage{
dist_k_ij(k, size)
}
\arguments{
\item{k}{kth indexes}

\item{size}{value of size attribute of the dist object}
}
\value{
ij indexes as 2*n matrix where n is length of k vector
}
\description{
Convert kth indexes to ij indexes of a dist object
}
