% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extractors.R
\name{dist_k_ij_}
\alias{dist_k_ij_}
\title{Convert kth index to ij index}
\usage{
dist_k_ij_(k, size)
}
\arguments{
\item{k}{kth index}

\item{size}{value of size attribute of the dist object}
}
\value{
ij index as a length two integer vector
}
\description{
Convert kth index to ij index of a dist object
}
