% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extractors.R
\name{dist_subset}
\alias{dist_subset}
\title{dist_subset}
\usage{
dist_subset(x, subset, ...)
}
\arguments{
\item{x}{dist object}

\item{subset}{index of the subset. This has to be unique.}

\item{...}{additional arguments}
}
\value{
returns a dist subset
}
\description{
Compute subset faster than regular `[[` on a dist object. This
  is from \pkg{proxy} package (not exported by proxy).
}
