% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_extractors.R
\name{distExtractPair}
\alias{distExtractPair}
\title{distExtractPair}
\usage{
distExtractPair(object, index1, index2)
}
\arguments{
\item{object}{dist object}

\item{index1}{index1}

\item{index2}{index2}
}
\value{
distance corresponding to indexes
}
\description{
extract distance between a pair of indexes
}
