% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{disto_bigdist}
\alias{disto_bigdist}
\title{Constructior of disto with bigdist backend}
\usage{
disto_bigdist(arguments)
}
\arguments{
\item{arguments}{to construct disto object}
}
\value{
returns a list
}
\description{
Constructior of disto with bigdist backend
}
\details{
to be used by disto constructor function
}
