% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{disto_dist}
\alias{disto_dist}
\title{Constructior of disto with dist backend}
\usage{
disto_dist(arguments)
}
\arguments{
\item{arguments}{to construct disto object}
}
\value{
returns a list
}
\description{
Constructior of disto with dist backend
}
\details{
to be used by disto constructor function
}
