% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_utils.R
\name{eps_k}
\alias{eps_k}
\title{Distance corresponding to kth neighbor}
\usage{
eps_k(x, k, ...)
}
\arguments{
\item{x}{Disto object}

\item{k}{A positive integer}

\item{...}{Arguments to 'dapply'. Should be among: subset, nproc, progress}
}
\value{
A vector of distances
}
\description{
Distance corresponding to kth neighbor
}
