% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{names.disto}
\alias{names.disto}
\title{Get names/labels}
\usage{
\method{names}{disto}(x)
}
\arguments{
\item{x}{disto object}
}
\value{
A character vector
}
\description{
Get names/labels of the underlying distance storing backend
}
\examples{
temp <- stats::dist(iris[,1:4])
dio <- disto(objectname = "temp")
dio
names(dio) <- paste0("a", 1:150)
}
