% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_utils.R
\name{nn}
\alias{nn}
\title{Nearest neighbors and distances}
\usage{
nn(distoObject, k, r, ...)
}
\arguments{
\item{distoObject}{Object of class 'disto'}

\item{k}{Number of nearest neighbors}

\item{r}{Radius for nearest neighbors}

\item{...}{Additional arguments for \code{\link{dapply}}}
}
\value{
A list with size of 'distObject' or length of the subset argument
  provided in dotdotdot. Each element of the list is a list with two
  elements: One is indexes (integer vector), other is distances (numeric
  vectors)
}
\description{
Obtain index of nearest neighbors and distances from a disto
  handle. k nearest or fixed radius neighbors are supported
}
\details{
Exactly one among k or r has to be provided
}
\examples{
do  <- dist(iris[, 1:4])
dio <- disto(objectname = "do")
nn(dio, k = 5)
nn(dio, r = 3)
}
