% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_utils.R
\name{nn_k}
\alias{nn_k}
\title{k Nearest neighbors}
\usage{
nn_k(vec, index, k)
}
\arguments{
\item{vec}{Vector of distances}

\item{index}{dummy to facilitate dapply}

\item{k}{Number of nearest neighbors}
}
\description{
k Nearest neighbors from a vector of distances
}
