% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_utils.R
\name{nn_r}
\alias{nn_r}
\title{Fixed radius Nearest neighbors}
\usage{
nn_r(vec, index, r)
}
\arguments{
\item{vec}{Vector of distances}

\item{index}{dummy to facilitate dapply}

\item{r}{Radius for nearest neighbors}
}
\description{
Fixed radius Nearest neighbors from a vector of distances
}
