% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{plot.disto}
\alias{plot.disto}
\title{Plot a disto object}
\usage{
\method{plot}{disto}(x, ...)
}
\arguments{
\item{x}{object of class disto}

\item{...}{Currently unused}
}
\value{
Plot output as side effect
}
\description{
Density plot a disto object
}
\examples{
temp <- stats::dist(iris[,1:4])
dio  <- disto(objectname = "temp")
plot(dio)
}
