% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{print.disto}
\alias{print.disto}
\title{Print method for dist class}
\usage{
\method{print}{disto}(x, ...)
}
\arguments{
\item{x}{object of class disto}

\item{...}{currently not in use}
}
\value{
invisible NULL. Function writes backend type and size to terminal as
  a message.
}
\description{
Print method for dist class
}
\examples{
temp <- stats::dist(iris[,1:4])
dio   <- disto(objectname = "temp")
print(dio)
}
