% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_utils.R
\name{print.nn}
\alias{print.nn}
\title{Print method for class 'nn'}
\usage{
\method{print}{nn}(x, ...)
}
\arguments{
\item{x}{Object of class 'nn'}

\item{...}{stub}
}
\value{
Returns the input invisibly besides printing on the screen
}
\description{
Print method for class 'nn'
}
