% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{size}
\alias{size}
\title{Obtain size of the disto object}
\usage{
size(disto, ...)
}
\arguments{
\item{disto}{object of class disto}

\item{...}{currently not in use}
}
\value{
Integer vector of length 1
}
\description{
Obtain size of the disto object
}
\examples{
temp <- stats::dist(iris[,1:4])
dio   <- disto(objectname = "temp")
size(dio)
}
