% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{summary.disto}
\alias{summary.disto}
\title{Summary method for dist class}
\usage{
\method{summary}{disto}(object, ...)
}
\arguments{
\item{object}{object of class disto}

\item{...}{currently not in use}
}
\value{
invisibly returns the tidy output of summary as a dataframe.
}
\description{
Summary method for dist class
}
\examples{
temp <- stats::dist(iris[,1:4])
dio   <- disto(objectname = "temp")
dio
summary(dio)
}
