% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disto_methods.R
\name{`names<-.disto``}
\alias{`names<-.disto``}
\alias{names<-.disto}
\title{Set names/labels}
\usage{
\method{names}{disto}(x) <- value
}
\arguments{
\item{x}{disto object}

\item{value}{A character vector}
}
\value{
invisible disto object
}
\description{
Set names/labels of the underlying distance storing backend
}
\examples{
temp <- stats::dist(iris[,1:4])
dio <- disto(objectname = "temp")
dio
names(dio) <- paste0("a", 1:150)
}
